Updated 08/2025 INSTALL.MD
I recently took on a MISP2Sentinel Integration for a client. Due to updates in the Azure portal, and steps that were documented in other sections or missing completely, it was a bit difficult to get it going. After troubleshooting and reading the code, I edited the document based on how it's exactly downloaded from this repo. 

I hope the changes were good, there were also some changes to just sections and wording to assist with walking through the setup.

Thank you!